moved confetti.js into scripts
