+++ src/scripts/confetti.js
@@ -225,4 +225,3 @@ export const stop = () => {
         confetti.stop()
     }, 3000); // 5000 is time that after 5 second stop the confetti ( 5000 = 5 sec)
 };
-
